Added validation for categories deletion

+++ src/controllers/category.js
@@ -1,4 +1,4 @@
-const { create, findAll,findById,updateById,deleteById,findByName } = require("../services/category");
+const { create, findAll,findById,updateById,deleteById,findByName,findAllExpenses } = require("../services/category");
 
 
 
@@ -26,7 +26,6 @@ exports.getCategories = async function (request, response) {
 	const categories = await findAll(id);
 	response.status(200).json(categories);
 };
-
 exports.getCategory = async function (request, response) {
 	const { id } = request.params;
 	const category = await findById(id);
@@ -47,13 +46,18 @@ exports.deleteCategory = async function (request, response) {
 	const currentCategory = await findById(id);
 
 	if (currentCategory) {
-		await deleteById(id);
-		response.status(204).end();
-	}
-	else {
+		const expenses = await findAllExpenses(id)
+		if (expenses == 0){
+			await deleteById(id);
+			response.status(204).end()
+			} else {
+				response.status(400).json({
+				message: "Cannot delete a category with expenses created"
+			});
+			}
+		} else {
 		response.status(400).json({
 			message: "Category does not exist"
 		});
 	};
-
 };
